Schistoxpkg.jl

A package to run an individual based model of a schistosomiasis outbreak based on original code from this paper. Generally people uptake larvae based on a contact rate defined by their age, along with some predisposition which is chosen from a gamma distribution with mean 1, but some specified level of variance.

All parameters are stored in the parameters.jl file in the src folder.

The model has a parameter which defines the time step that we take forward each time. Several functions are then called each time step which simulate the run of the outbreak. This is repeated until we reach a specified number of steps, usually corresponding to stepping forward a chosen number of years into the future.

The standard approach is given by the following set of processes which all have their own function to execute them. First, load required packages and include the parameters file which stores the parameters for the model.

using Schistoxpkg
using Random
using JLD
using Distributions
using PyPlot
include("parameters.jl")

After this, we must initialize the pars struct, which will be used to store the parameters.

pars = Parameters(N, time_step, N_communities, community_probs, community_contact_rate,
                  density_dependent_fecundity, average_worm_lifespan,
                  max_age, initial_worms, initial_miracidia, 
                  initial_miracidia_days, init_env_cercariae,
                  worm_stages, contact_rate, max_fecundity, age_contact_rates,
                  ages_for_contacts, contact_rate_by_age_array, mda_adherence, 
                  mda_access,  female_factor, male_factor, miracidia_maturity,
                  birth_rate, human_cercariae_prop, predis_aggregation, 
                  cercariae_survival, miracidia_survival,
                  death_prob_by_age, ages_for_death, r, 
                  vaccine_effectiveness, drug_effectiveness,
                  spec_ages, ages_per_index, record_frequency, 
                  use_kato_katz, kato_katz_par, heavy_burden_threshold,
                  rate_acquired_immunity, M0, human_larvae_maturity_time)

pars = make_age_contact_rate_array(pars, scenario, [],[]);

We then counstruct the initial host population and initialize the environment cercariae and mircacidia larvae. We also age the population forward for a certain number of steps (here 20,000), so that there are no individuals in the population whose death age is lower than their actual age, and the death rate dynamics have time to achieve a desired age distribution. We also update the age based contact rates according to the new ages in the population.

humans, miracidia, cercariae = create_population_specified_ages(pars)
humans = generate_ages_and_deaths(20000, humans, pars)
humans = update_contact_rate(humans,  pars)

mda_info = []

vaccine_info = []

Each time step we advance the time of the simulation by the length of the time step and also add this time step to the age of each individual. There is a chosen period at which contact rates are updated for each individual, where we check if someone has aged into a different age bracket, resulting if their level of contact has changed.

We then calculate the total number of worms within each individual and the number of pairs of worms a person has. These numbers are used to calculate how many eggs someone will produce. The number of eggs is chosen from a poisson distribution with mean equal to the number of worm pairs multiplied by the max fecundity parameter and then multiplied by an exponential function which calculates the density dependent reduction in eggs produced, λ wp exp(-wp z). We then kill the worms within human hosts at a given rate, which is based on average worm lifespan.

Eggs are then hatched into the environment, with egg release dependent on the age specific contact rate of each individual. Humans are given an age of death when they are born, which is based on some chosen death rates for each age group. We check each time step if anyone has outlived their age of death and if they have, they are then removed from the population. Cercariae from the environment are then uptaken to each surviving individual based on their predisposition and contact rate. These immediately become worms within the human host.

We then perform any interventions which are due to take place at this point in time after which we will cull the miracidia and cercariae in the environment by a chosen percentage. After this we will add births to the population which occur at some specified rate.

A version of this is done by with the following function:

number_years = 200
num_time_steps = trunc(Int, 365*number_years / time_step)
humans, miracidia, cercariae, record = 
update_env_no_births_deaths_human_larvae(num_time_steps, humans,  miracidia, 
                                         cercariae, pars, mda_info, vaccine_info)

There are other versions of this basic approach, where we don't age the population or include births and deaths and also where the population is aged but every death is simply matched with a birth, resulting in the population being kept constant.

Schistoxpkg.administer_drug

Schistoxpkg.administer_vaccine

Schistoxpkg.birth_of_human

Schistoxpkg.calculate_worm_pairs

Schistoxpkg.cercariae_death!

Schistoxpkg.cercariae_uptake!

Schistoxpkg.cercariae_uptake_with_human_larvae!

Schistoxpkg.create_population

Schistoxpkg.create_population_specified_ages

Schistoxpkg.death_of_human

Schistoxpkg.egg_production!

Schistoxpkg.egg_production_increasing!

Schistoxpkg.human_larvae_maturity

Schistoxpkg.mda

Schistoxpkg.miracidia_death!

Schistoxpkg.miracidia_production!

Schistoxpkg.update_contact_rate

Schistoxpkg.worm_maturity!

Schistoxpkg.administer_drug — Method

administer_drug(humans, indices, drug_effectiveness)

administer mda drugs to chosen individuals in the population. If they adhere to the drugs, then they reduce male and female worms with a given efficacy alongside removing eggs

Schistoxpkg.administer_vaccine — Method

administer_vaccine(humans, indices, vaccine_effectiveness, vaccine_duration)

administer vaccine to chosen individuals in the population. reduce male and female worms with a given efficacy alongside removing eggs and adding to their vaccine status signifying that they will have increased immunity for a chosen period of time

Schistoxpkg.birth_of_human — Method

birth_of_human(humans, pars)

add an individual to the population

Schistoxpkg.calculate_worm_pairs — Method

calculate_worm_pairs(female_worms, male_worms)

calculate how many pairs of worms there are in each human host

Schistoxpkg.cercariae_death! — Method

cercariae_death!(miracidia, pars)

Kill a chosen proportion of cercariae in the environment governed by the cercariae_survival parameter in the pars struct

Schistoxpkg.cercariae_uptake! — Method

cercariae_uptake(humans, cercariae, miracidia, pars)

uptake cercariae into humans, whilst updating cercariae with miracidia. Uptaken cercariae immediately become worms in this formulation

Schistoxpkg.cercariae_uptake_with_human_larvae! — Method

cercariae_uptake_human_larvae!(humans, cercariae, miracidia, pars)

uptake cercariae into humans, whilst updating cercariae with miracidia. Uptaken cercariae become larvae within humans, rather than immmediately into worms with this function.

Schistoxpkg.create_population — Method

create_population

This will create the initial human population with randomly chosen age, and gender. Predisposition is taken to be gamma distributed There is also a male and female adjustment to predisposition adjusting for gender specific behaviour In addition to this, it will create the initial miracidia environment vector

Schistoxpkg.create_population_specified_ages — Method

create_population_specified_ages(pars)

This will create the initial human population with an age distribution specified by the spec_ages variable Predisposition is taken to be gamma distributed. There is also a male and female adjustment to predisposition adjusting for gender specific behaviour In addition to this, it will create the initial miracidia environment vector

Schistoxpkg.death_of_human — Method

death_of_human(humans)

remove individuals from the population whose age is greater than their death age

Schistoxpkg.egg_production! — Method

egg_production!(humans, pars)

function to produce eggs for individuals, dependent on how many worms they have and the max fecundity and density dependent fecundity of the population

Schistoxpkg.egg_production_increasing! — Method

egg_production_increasing!(humans, pars)

function to produce eggs for individuals, dependent on how many worms they have and the max fecundity and density dependent fecundity of the population

Schistoxpkg.human_larvae_maturity — Method

human_larvae_maturity(humans, pars)

This will mature the human larvae into worms after a chosen number of days, which is specified by the humanlarvaematurity_time parameter in the pars struct

Schistoxpkg.mda — Method

mda(humans, mda_coverage, min_age_mda, max_age_mda, mda_effectiveness, mda_gender)

administer mda in the population. This includes choosing individuals between specified ages, having a certain level of coverage and taking access and adherence into consideration

Schistoxpkg.miracidia_death! — Method

miracidia_death!(miracidia, pars)

Kill a chosen proportion of miracidia in the environment governed by the miracidia_survival parameter in the pars struct

Schistoxpkg.miracidia_production! — Method

miracidia_production!(humans)

release eggs from individuals into the environment as miracidia. Release is relative to the contact rate with the environment for each individual.

Schistoxpkg.update_contact_rate — Method

update_contact_rate(humans, pars)

function to update the contact rate of individuals in the population. This is necessary as over time when people age, they will move through different age groups which have different contact rates

Schistoxpkg.worm_maturity! — Method

worm_maturity!(humans, pars)

function to kill worms within human hosts, and if there is more than one stage for worm life, to update how many worms are in each stage