Schistoxpkg.jl

A package to run an individual based mode of a schistosomiasis outbreak.

All parameters are stored in the parameters.jl file in the src folder.

To run to equilibrium, a few functions need to be fun. After defining parameters we have to run "contactratesbyage" = makeagecontactratearray(maxage, scenario, [], []); to generate the contact rate by age array. After this, we have to run:

ages , deathages, gender, predisposition, community, humancercariae, eggs, vacstatus, treated, femaleworms, maleworms, agecontactrate, vaccinated, envmiracidia, adherence, access = createpopulationspecifiedages(N, Ncommunities, communityprobs, initialworms, contactratesbyage, wormstages, femalefactor, malefactor,initialmiracidia, initialmiracidiadays, predisaggregation, predisweight, timestep, specages, agesperindex, deathprobbyage, agesfordeaths, mdaadherence, mdaaccess)

deathages = [] for i in 1:N push!(deathages, getdeathage(deathprobbyage, agesfordeaths)) end mean(deathages) ages, deathages = generateagesanddeaths(20000, ages, deathages, deathprobbyage, agesfordeaths)

agesequ, genderequ, predispositionequ, humancercariaeequ, eggsequ, vacstatusequ, treatedequ, femalewormsequ, malewormsequ, vaccinatedequ, envmiracidiaequ, envcercariaeequ, recordhigh = updateenvtoequilibrium(7600, copy(ages), copy(humancercariae), copy(femaleworms), copy(maleworms), copy(community), communitycontactrate, timestep, averagewormlifespan, copy(eggs), maxfecundity, r, wormstages, copy(vacstatus), copy(gender), predisaggregation, copy(predisposition), copy(treated), vaccineeffectiveness, densitydependentfecundity, copy(vaccinated), copy(envmiracidia), copy(envcercariae), contactrate, envcercariaesurvivalprop, envmiracidiasurvivalprop, femalefactor, malefactor, contactratesbyage, recordfrequency, copy(agecontactrate),humancercariaeprop, miracidiamaturitytime)

Schistoxpkg.create_population

Schistoxpkg.create_population_specified_ages

Schistoxpkg.create_population — Method

create_population

This will create the initial human population with randomly chosen age, and gender.
 Predisposition is taken to be gamma distributed.
 There is also a male and female adjustment to predisposition adjusting for gender specific behaviour
     In addition to this, it will create the initial miracidia environment vector

Schistoxpkg.create_population_specified_ages — Method

create_population_specified_ages

This will create the initial human population with an age distribution specified by the spec_ages variable Predisposition is taken to be gamma distributed. There is also a male and female adjustment to predisposition adjusting for gender specific behaviour In addition to this, it will create the initial miracidia environment vector