Functions

Functions included in the Schistoxpkg.

Schistoxpkg.Human — Type

Human struct

This struct contains the information for the vaccine, storing the coverage, minimum and maximum age targeted, gender, drug efficacy and the time for the vaccine to be done along with how long the vaccine provides protection for

Schistoxpkg.mda_information — Type

mda_information struct

This struct contains the information for the mda, storing the coverage, minimum and maximum age targeted, gender, drug efficacy and the time for the mda to be done

Schistoxpkg.out — Type

out struct

This struct contains the different outputs we are interested in recording. This is the overall population burden, with categories for low, moderate and heavy burdens, along with separate categories for the school age children and adults. Along with these, the time of each result is recorded, so we can subsequently see the prevalence of the otubreak over time.

Schistoxpkg.vaccine_information — Type

vaccine_information struct

This struct contains the information for the vaccine, storing the coverage, minimum and maximum age targeted, gender, drug efficacy and the time for the vaccine to be done along with how long the vaccine provides protection for

Schistoxpkg.administer_drug — Method

administer_drug(humans, indices, drug_effectiveness)

administer mda drugs to chosen individuals in the population. If they adhere to the drugs, then they reduce male and female worms with a given efficacy alongside removing eggs

Schistoxpkg.administer_vaccine — Method

administer_vaccine(humans, indices, vaccine_effectiveness, vaccine_duration)

administer vaccine to chosen individuals in the population. reduce male and female worms with a given efficacy alongside removing eggs and adding to their vaccine status signifying that they will have increased immunity for a chosen period of time

Schistoxpkg.birth_of_human — Method

birth_of_human(humans, pars)

add an individual to the population

Schistoxpkg.calculate_worm_pairs — Method

calculate_worm_pairs(female_worms, male_worms)

calculate how many pairs of worms there are in each human host

Schistoxpkg.cercariae_death! — Method

cercariae_death!(miracidia, pars)

Kill a chosen proportion of cercariae in the environment governed by the cercariae_survival parameter in the pars struct

Schistoxpkg.cercariae_uptake! — Method

cercariae_uptake(humans, cercariae, miracidia, pars)

uptake cercariae into humans, whilst updating cercariae with miracidia. Uptaken cercariae immediately become worms in this formulation

Schistoxpkg.cercariae_uptake_with_human_larvae! — Method

cercariae_uptake_human_larvae!(humans, cercariae, miracidia, pars)

uptake cercariae into humans, whilst updating cercariae with miracidia. Uptaken cercariae become larvae within humans, rather than immmediately into worms with this function.

Schistoxpkg.count_eggs — Method

count_eggs(humans)

count the total number of eggs in the human population

Schistoxpkg.create_mda — Method

create_mda(pre_SAC_prop, SAC_prop, adult_prop, first_mda_time,
        last_mda_time, regularity, pre_SAC_gender, SAC_gender, adult_gender, mda_effectiveness)

function to create a set of mda's which will be performed regularly firstmdatime specifies when this will first occur in years, lastmdatime is the final mda in this block regularity is how often to perform the mda in years. specify the proportion of pre SAC, SAC and adults at each of these time points also specify genders for these differect age groups, along with the effectiveness of mda

Schistoxpkg.create_population — Method

create_population

This will create the initial human population with randomly chosen age, and gender. Predisposition is taken to be gamma distributed There is also a male and female adjustment to predisposition adjusting for gender specific behaviour In addition to this, it will create the initial miracidia environment vector

Schistoxpkg.create_population_specified_ages — Method

create_population_specified_ages(pars)

This will create the initial human population with an age distribution specified by the spec_ages variable Predisposition is taken to be gamma distributed. There is also a male and female adjustment to predisposition adjusting for gender specific behaviour In addition to this, it will create the initial miracidia environment vector

Schistoxpkg.death_of_human — Method

death_of_human(humans)

remove individuals from the population whose age is greater than their death age

Schistoxpkg.egg_production! — Method

egg_production!(humans, pars)

function to produce eggs for individuals, dependent on how many worms they have and the max fecundity and density dependent fecundity of the population

Schistoxpkg.egg_production_increasing! — Method

egg_production_increasing!(humans, pars)

function to produce eggs for individuals, dependent on how many worms they have and the max fecundity and density dependent fecundity of the population

Schistoxpkg.get_prevalences! — Method

get_prevalences!(humans, time, pars)

calculate the desired prevalences in the human population.

Schistoxpkg.human_larvae_maturity — Method

human_larvae_maturity(humans, pars)

This will mature the human larvae into worms after a chosen number of days, which is specified by the humanlarvaematurity_time parameter in the pars struct

Schistoxpkg.kato_katz — Method

kato_katz(eggs, gamma_k)

calculate number of eggs using kato katz method. Gammak is a gamma distribution with shape and scale defined by pars.katokatz_par

Schistoxpkg.mda — Method

mda(humans, mda_coverage, min_age_mda, max_age_mda, mda_effectiveness, mda_gender)

administer mda in the population. This includes choosing individuals between specified ages, having a certain level of coverage and taking access and adherence into consideration

Schistoxpkg.miracidia_death! — Method

miracidia_death!(miracidia, pars)

Kill a chosen proportion of miracidia in the environment governed by the miracidia_survival parameter in the pars struct

Schistoxpkg.miracidia_production! — Method

miracidia_production!(humans)

release eggs from individuals into the environment as miracidia. Release is relative to the contact rate with the environment for each individual.

Schistoxpkg.update_contact_rate — Method

update_contact_rate(humans, pars)

function to update the contact rate of individuals in the population. This is necessary as over time when people age, they will move through different age groups which have different contact rates

Schistoxpkg.update_mda — Method

update_mda(mda_info, mda_round)

update when the next mda will take place

Schistoxpkg.vac_decay! — Method

vac_decay!(humans)

decrease vaccination status for each person by 1 each day

Schistoxpkg.worm_maturity! — Method

worm_maturity!(humans, pars)

function to kill worms within human hosts, and if there is more than one stage for worm life, to update how many worms are in each stage